Test #5
Still not ready to give this a more definitive title until i fix some
core things, but in the meantime…

- Fixed some issues with the moderator system, should work just fine now
- Restructured a bunch of code to help with readability and management,
slicing up both Pelican and Sessions.  More work is to be done on this
front…
- Added an initial version for “Prompt”, a Session-based system
designed to help you provide inline options to users and respond to
them in an elegant way.  Currently this will only work for one inline
message at a time per Session, but ill provide containers at a later
point so different kinds of inline options can always be stored and
responded to.