Death Pt.4
- *NEW* SessionTag - Used as both a means for delegates to identify the
Session it belongs to, and a callback portal to make Telegram API
requests and basic changes to the Session it belongs to, through
Pelican.
- *NEW* SessionEvent - Properly made this time, it acts as a simple
container thats created by a SessionTag when a delegate requests
changes.
- *NEW* Flood R2 - Flood controls are now more customisable and let you
define what update types a flood monitor reacts to, and what happens
when it reaches the required update hits.
- *NEW* Timeout R2 - Timeouts are now contained in a type that utilises
the schedule to bounce back removal requests.  Set it how you like and
watch it roll!

## Changes
- ChatSession now has Flood and Timeout controls
- SessionBuilder can now remove Sessions using a SessionTag.

## Up Next
- Revised Moderation system and developer controls
- Revised and working Prompt system
- Revised and working file uploads